+++ lib/mixins.dart
@@ -1,6 +1,13 @@
 library mixins;
 
+// external packages
+export 'package:intl/intl.dart';
+export 'package:path_provider/path_provider.dart';
+export 'package:stack_trace/stack_trace.dart';
+
+export 'packages/shimmer.dart';
 export 'src/extension.dart';
+export 'src/formatter.dart';
 export 'src/helper.dart';
 export 'src/lipsum.dart';
 export 'src/log.dart';