+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2
 
 * second release
+
+## 0.0.3
+
+* add input formatter
+* add some helper method